Move game state smachs to separate files in agent/src
