+++ edante/agent/src/agent/__init__.py
@@ -1,1 +1,2 @@
-from agent import *
+from playing_sm import PlayingSM
+from finished_sm import FinishedSM
